A37: properties: Drop persist.data.target
We have it set to dpm3 originally, but some users reported mobile data
will not work properly with this config. They reported that dpm1 is better
for them as it made mobile data work.

Dropping it will fix mobile data for all as I assume it will automatically
select the best value for the user.

Reported-by: Leviathan Melvilli <skmalladi30@gmail.com>
Signed-off-by: Ricky Cheung <rcheung844@gmail.com>